Refactor project structure: Organize files into koji and task packages
