+++ src/main/java/koji/KojiException.java
@@ -1,3 +1,5 @@
+package koji;
+
 public class KojiException extends Exception {
     public KojiException(String message) {
         super(message);
