+++ src/main/java/task/Deadline.java
@@ -1,3 +1,5 @@
+package task;
+
 public class Deadline extends Task {
     protected String by;
