+++ src/main/java/task/TaskList.java
@@ -1,3 +1,5 @@
+package task;
+
 import java.util.ArrayList;
 
 public class TaskList {
