+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ____________________________________________________________
- Hello! I'm Koji
+ Hello! I'm koji.Koji
  What can I do for you?
 ____________________________________________________________
 ____________________________________________________________
